feat: 카테고리 이미지 컴포넌트 함수형으로 변경
