--- src/components/CategoryImage.tsx
@@ -1,32 +1,0 @@
-import React, { Component } from 'react';
-import styles from './CategoryImage.module.css';
-import { FoodCategory } from '../types/restaurantInfo';
-import KOREAN_IMAGE from '../assets/category-korean.png';
-import JAPANESE_IMAGE from '../assets/category-japanese.png';
-import CHINESE_IMAGE from '../assets/category-chinese.png';
-import ASIAN_IMAGE from '../assets/category-asian.png';
-import WESTERN_IMAGE from '../assets/category-western.png';
-import ETC_IMAGE from '../assets/category-etc.png';
-
-const CATEGORY_IMAGE: Record<string, string> = {
-  한식: KOREAN_IMAGE,
-  일식: JAPANESE_IMAGE,
-  중식: CHINESE_IMAGE,
-  아시안: ASIAN_IMAGE,
-  양식: WESTERN_IMAGE,
-  기타: ETC_IMAGE,
-};
-
-class CategoryImage extends Component<{ category: FoodCategory }> {
-  render() {
-    const { category } = this.props;
-
-    return (
-      <div className={styles.background} title={category}>
-        <img src={CATEGORY_IMAGE[category]} alt={category} />
-      </div>
-    );
-  }
-}
-
-export default CategoryImage;
