+++ src/components/CategoryImage/CategoryImage.module.css
