+++ src/components/CategoryImage/assets/category-asian.png
