+++ src/components/CategoryImage/assets/category-chinese.png
