+++ src/components/CategoryImage/assets/category-japanese.png
