+++ src/components/CategoryImage/assets/category-korean.png
