+++ src/components/CategoryImage/assets/category-western.png
